split generation into own class
