+++ lib/index.js
@@ -1,17 +1,16 @@
 var Grid = require('./grid');
+var Generator = require('./generator');
 var DisplayHandler = require('./display_handler');
 
 var canvas = document.getElementById('game');
 var context = canvas.getContext('2d');
 
 var grid = new Grid({width: 56, height: 40});
+var generator = new Generator(grid);
 var displayHandler = new DisplayHandler({context: context, grid: grid});
 
-grid.generateRooms(50, 3, 11);
-grid.generateMazes();
-//grid.getCell(1,1).setValue(2);
-//grid.createRoom({x: 2, y: 2, width: 2, height: 2});
-//grid.createRoom({x: 5, y: 2, width: 2, height: 2});
+generator.generateRooms(50, 3, 11);
+//generator.generateMazes();
 
 requestAnimationFrame(function drawLoop() {
     displayHandler.draw();
